Update patch set 9
Patch Set 9:

So..after wasting lot of time debugging, seems like it's setuptools issue in python2 where the tests are not discovered.

https://github.com/pypa/setuptools/pull/1187/commits/84093b78ec61ad47a2a0dea9f1be8d94fa0d485e

Anyway the tests have a dependency with ansible-runner for which the package does not exist in RDO. Can we go ahead and land this?

Patch-set: 9
Reviewer: Gerrit User 415 <415@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0